+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ### 🐛 Fixes
 
 - [8962](https://github.com/vegaprotocol/vega/issues/8962) - Refreshed pegged iceberg orders remain tracked as pegged orders.
+- [8968](https://github.com/vegaprotocol/vega/issues/8968) - Fix missing parent reference on checkpoint restore.
 
 ## 0.72.6